requirements.txt
torch>=1.8.0
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.0
gensim>=4.0.0
jieba>=0.42.1
matplotlib>=3.3.0
pyyaml>=5.4.0
tqdm>=4.50.0
accelerate>=0.20.0 